Refactor comparison
