+++ burn-wgpu/src/kernel/comparison/mod.rs
@@ -1,0 +1,7 @@
+mod base;
+mod binary;
+mod elem;
+
+pub use base::*;
+pub use binary::*;
+pub use elem::*;
