+++ burn-wgpu/src/kernel/mod.rs
@@ -2,7 +2,6 @@ mod base;
 mod binary_elemwise;
 mod cat;
 mod comparison;
-mod comparison_elem;
 mod index;
 mod mask;
 mod matmul;
@@ -20,7 +19,6 @@ pub use unary_scalar::*;
 
 pub(crate) use cat::*;
 pub(crate) use comparison::*;
-pub(crate) use comparison_elem::*;
 pub(crate) use index::*;
 pub(crate) use mask::*;
 pub(crate) use reduction::*;
